Added tag support

+++ simplecd.sh
@@ -6,7 +6,7 @@
 PATH=$PATH:/bin:/usr/bin:/usr/sbin:/usr/local/bin
 
 
-# Functions ####################################################################
+# Function #################################################################
 
 shutdown () {
   echo $1
@@ -16,7 +16,7 @@ shutdown () {
   exit 0
 }
 
-abort () { 
+abort () {
   echo $1
   echo ""
   mail_log "failure"
@@ -35,7 +35,7 @@ mail_log () {
     echo "$LOG" > $WORKINGDIR/log.$HASH
     while read MR; do
       echo "Mailing log of this run to $MR..."
-      mail -aFrom:simplecd@`hostname --fqdn` -s "[simplecd][$1] $REPO - $BRANCH" $MR < $WORKINGDIR/log.$HASH
+      mail -aFrom:simplecd@`hostname --fqdn` -s "[simplecd][$1] $REPO - $SOURCE" $MR < $WORKINGDIR/log.$HASH
     done < $REPODIR/_simplecd/logreceivers.txt
   fi
 }
@@ -47,7 +47,7 @@ run_project_script () {
   log "Output of project's $1 script:
 #######################################"
   echo ""
-  OUTPUT=`$REPODIR/_simplecd/$1 $REPODIR $BRANCH 2>&1`
+  OUTPUT=`$REPODIR/_simplecd/$1 $MODE $REPODIR $CHECKOUTSOURCE 2>&1`
   STATUS=$?
   echo "$OUTPUT"
   log "$OUTPUT"
@@ -64,8 +64,22 @@ run_project_script () {
 
 # Main routine #################################################################
 
-REPO=$1
-BRANCH=$2
+MODE=$1 # "branch" or "tag"
+SOURCE=$2
+REPO=$3
+
+if [ "$MODE" = "" ]; then
+  abort "Missing parameter MODE. Aborting..."
+fi
+
+if [ "$REPO" = "" ]; then
+  abort "Missing parameter REPO. Aborting..."
+fi
+
+if [ "$SOURCE" = "" ]; then
+  abort "Missing parameter SOURCE. Aborting..."
+fi
+
 
 if [ -x /sbin/md5 ]; then
   MD5BIN=/sbin/md5
@@ -73,7 +87,7 @@ else
   MD5BIN=/usr/bin/md5sum
 fi
 
-HASH=`echo "$0 $REPO $BRANCH" | $MD5BIN | cut -d" " -f1`
+HASH=`echo "$0 $MODE $REPO $SOURCE" | $MD5BIN | cut -d" " -f1`
 WORKINGDIR=/var/tmp/simplecd
 PROJECTSDIR=$WORKINGDIR/projects
 REPODIR=/var/tmp/simplecd/projects/$HASH
@@ -82,36 +96,28 @@ CONTROLFILE=/var/tmp/simplecd/controlfile.$HASH
 # Did the user provide the parameter "reset"? In this case
 # we remove everything we know about the given repo/branch combination
 
-if [ "$3" = "reset" ]; then
-  echo "Resetting SimpleCD environment for repo $REPO, branch $BRANCH"
-  rm -f $WORKINGDIR/last_commit_id.$HASH
+if [ "$4" = "reset" ]; then
+  echo "Resetting SimpleCD environment for mode $MODE, repo $REPO, source $SOURCE"
+  if [ "$MODE" = "branch" ]; then
+    rm -f $WORKINGDIR/last_commit_id.$HASH
+  fi
+  if [ "$MODE" = "tag" ]; then
+    rm -f $WORKINGDIR/last_tag.$HASH
+  fi
   rm -f $WORKINGDIR/log.$HASH
   rm -rf $REPODIR
   rm -f $CONTROLFILE
   echo "done."
   exit 0
 fi
-URLPREFIX=$3
+URLPREFIX=$4
 
-# Is another process for this repo and branch running?
+# Is another process for this mode, repo and source running?
 
 if [ -f $CONTROLFILE ]; then
   echo "Because the control file $CONTROLFILE exists, I assume that another instance is still running. Aborting..."
   exit 1
 fi
-
-
-# Create control file so no other runs are started in parallel
-
-touch $CONTROLFILE
-
-
-# Let's go
-
-echo ""
-echo "Starting delivery of branch $BRANCH from repo $REPO, hash of this run is $HASH"
-echo ""
-log "Log for delivery of branch $BRANCH from repo $REPO, hash of this run was $HASH"
 
 # Prepare and check the environment
 
@@ -121,42 +127,72 @@ if [ ! -w $PROJECTSDIR ]; then
   abort "Cannot write to directory $PROJECTSDIR. Aborting..."
 fi
 
+# Create control file so no other runs are started in parallel
 
-# Check if a new commit id is in the remote repo
-
-LASTCOMMITID=`cat $WORKINGDIR/last_commit_id.$HASH 2> /dev/null`
-REMOTECOMMITID=`git ls-remote $REPO refs/heads/$BRANCH | cut -f1`
-
-if [ "$LASTCOMMITID" = "$REMOTECOMMITID" ]; then
-  echo "Remote commit id ($REMOTECOMMITID) has not changed since last run, won't deliver. Aborting..."
-  rm -f $CONTROLFILE
-  exit 0
-fi
+touch $CONTROLFILE
 
 
-# Clone the repo and checkout the branch
+# Let's go
+
+echo ""
+echo "Starting delivery of source $SOURCE from repo $REPO in mode $MODE, hash of this run is $HASH"
+echo ""
+log "Log for delivery of source $SOURCE from repo $REPO in mode $MODE, hash of this run was $HASH"
 
 rm -rf $REPODIR
 git clone $REPO $REPODIR 2>&1 | while IFS= read -r line;do echo " [GIT CLONE] $line";done
 cd $REPODIR
-git checkout $BRANCH 2>&1 | while IFS= read -r line;do echo " [GIT CHECKOUT] $line";done
+git fetch
 
 
-# Store the current commit id
+# Resolve source and check for new content
 
-CURRENTCOMMITID=`git log -n 1 refs/heads/$BRANCH --pretty=format:"%H"`
+if [ "$MODE" = "branch" ]; then
+  RESOLVEDSOURCE=refs/heads/$SOURCE
+  # Check if a new commit id is in the remote repo
+  LASTCOMMITID=`cat $WORKINGDIR/last_commit_id.$HASH 2> /dev/null`
+  REMOTECOMMITID=`git ls-remote $REPO $RESOLVEDSOURCE | cut -f1`
+  if [ "$LASTCOMMITID" = "$REMOTECOMMITID" ]; then
+    echo "Remote commit id ($REMOTECOMMITID) has not changed since last run, won't deliver. Aborting..."
+    rm -f $CONTROLFILE
+    exit 0
+  fi
+  CURRENTCOMMITID=`git log -n 1 $RESOLVEDSOURCE --pretty=format:"%H"`
+  echo $CURRENTCOMMITID > $WORKINGDIR/last_commit_id.$HASH
+  CHECKOUTSOURCE=$SOURCE
+fi
+if [ "$MODE" = "tag" ]; then
+  # Check if a new tag matching the pattern exists
+  LASTTAG=`cat $WORKINGDIR/last_tag.$HASH 2> /dev/null`
+  LASTEXISTINGTAG=`git tag -l $SOURCE | tail -n1`
+  if [ "$LASTTAG" = "$LASTEXISTINGTAG" ]; then
+    echo "No tag newer than '$LASTTAG' found, won't deliver. Aborting..."
+    rm -f $CONTROLFILE
+    exit 0
+  fi
+  CURRENTCOMMITID=$LASTEXISTINGTAG
+  echo $LASTEXISTINGTAG > $WORKINGDIR/last_tag.$HASH
+  RESOLVEDSOURCE=refs/tags/$LASTEXISTINGTAG
+  CHECKOUTSOURCE=$LASTEXISTINGTAG
+fi
 
-echo $CURRENTCOMMITID > $WORKINGDIR/last_commit_id.$HASH
+
+# Checkout the source
+
+git checkout $CHECKOUTSOURCE 2>&1 | while IFS= read -r line;do echo " [GIT CHECKOUT] $line";done
+
+
+# Create summary
 
 echo ""
 echo "This is what's going to be delivered:"
 SUMMARY="
  Repository: $REPO
-     Branch: $BRANCH
+     Source: $MODE $RESOLVEDSOURCE
      Commit: $URLPREFIX$CURRENTCOMMITID
-         by: `git log -n 1 refs/heads/$BRANCH --pretty=format:'%an'`
-         at: `git log -n 1 refs/heads/$BRANCH --pretty=format:'%aD'`
-        msg: `git log -n 1 refs/heads/$BRANCH --pretty=format:'%s'`"
+         by: `git log -n 1 $CURRENTCOMMITID --pretty=format:'%an'`
+         at: `git log -n 1 $CURRENTCOMMITID --pretty=format:'%aD'`
+        msg: `git log -n 1 $CURRENTCOMMITID --pretty=format:'%s'`"
 echo "$SUMMARY"
 log "$SUMMARY"
 echo ""
